requirements.txt
PyYAML>=6.0.2
Jinja2>=3.1.4
httpx>=0.27.2
# PNG rendering (optional but recommended)
playwright>=1.44.0
